Config: add --close-after-days option
So that you can specify more precisely control the timing of stages
that issues go through during the process:

  1. waiting (--stale-after-days)
  2. Comment on issue
  3. waiting  (--close-after-days)
  4. Close issue

Also, change the "stale" state to be dependent on the dat of the last
comment, not the last edit. Otherwise any edit to any comment will
keep an issue open. We only consider new comments to be important.